Replace main function in modules with mangled version of filaname
This is in preperation for multi-module builds, which means you cannot have
multiple main()'s floating around. When compiling to an actual binary though,
you still need a main function. To fix this, there is a handy shim that calls
the modules "main" function from a real C main function.