Create very basic semantic analysis pipeline
A big issue with the current codegeneration system is that it is tyightly
coupled with LLVM. In addition to the LLVM specific logic, there is
generic semantic information which could apply to all (future) backends.

To resolve this, a new, totally isolated semantic analysis pipeline is
being added, which will hopefully solve this.