Add ability to print value of function parameters in GDB:
To make this work we have to do this "hack" where you turn the function
parameters (which are registers) into stack allocated variables. Turns
out you cannot call "llvm.dbg.declare" on a register, hence why you
need to re-create the variable. Clang does this whether you are in debug
mode or not, whereas our version only does it in debug mode. This could
be modified create to work on both debug and non-debug versions, but
that is for another day (and might be needed anyways when/if mutable
parameters are added).

Note: When a breakpoint on a function is first hit, GDB will print the
function and its parameters, which now says some random (garbage) value.
It is only when you step into the function that the real value is set.