Handle variable use-before-declaration in the semantic layer:
Before, a variable "being defined" was whether or not you had an LLVM value ref
attached to the variable. Now that this kind of semantic logic is being moved to
the semantic layer (which has no concept of LLVM), things got a little messy.
This fixes that issue by creating a flag that the semantic layer controls, and
removes the need for the "allow_uninitialized" param in the search function.