Add very basic instrumentation for e2e tests:
Had to fix a few issues with how the LLVM contexts are being used. Turns out
the main modules where being created in the global LLVM context, which is bad
news. There also was a use-after-free bug with the semantic state scope since
it was not NULL'd afterwards (since it was not expecting to be reused).