Fix order of operations:
Previously, all expressions where evaluated right-to-left, and had no concept
of operator precedence. This fixes that, and implements proper operator
precedence. Operator precedence is different from C, and was just what I felt
made sence, and is subject to change.

Note: The only expressions which do not follow order of operations is the
power operator, which should evaulate from right to left, though this is good
enough for now.