Fix default value for variables being incorrectly assigned:
Previously the LLVM value was assigned to the global variables if it was
marked as export (which doesnt make sense). Now the default value is set
if the expression is constant. Now that we are assigning a default value
for globals, we dont have to add a store instruction for assigning the
value, since it will aready be assigned (unless it is not const expr).