Add basic embed program:
One of the big issues with the e2e tests is that they are just a bunch of
strings, so running specific tests require extracting them to files. Also,
there is no syntax high-lighting, which makes it annoying to look at. By
"compiling" the .sk and .sk.ll files into a C fixture, we can still have
the files stored in text format, with runtime IO overhead in e2e tests (we
still have to compile the initial files, but after that it shouldn't require
extra IO).