Drop support for LLVM 12:
The LLVM 13 C API was the first to get access to new pass manager:

https://reviews.llvm.org/D102136

So by adding optimization support in 538b442a, I made made LLVM 12 unsupported.
This is fine, since the LLVM version should be specific anyways. The only issue
is that support for LLVM 13 in ubuntu is lacking, but will get better when
22.04 is released.