Allow for dereference operator to be used without a space:
* Remove ambiguities in is_unary_oper function:

Since certain tokens (dash and star) can be used in both a binary and unary
context, the parsing of these token types are context sensitive. In certain
cases, we are expecting a unary operator, and as such, should except it. In
other cases, we are expecting a binary operator, and cannot have a unary
operator. To fix this, we just make 2 functions: one that only allows for the
exact ExprType, and another that allows for the context-sensitive TokenType.