Add basic name mangling for exported funcs/variables:
To allow for exporting and importing of symbols in modules, we need a way to
namespace things when linking together. In C, we will get conflicts if 2
symbols with the same name exist in the same program, even of they are not
used in the same module. To fix this, we prefix exported variables and
functions with "module_name.", the period being a character which cannot
appear in normal identifiers. The issue still persists when we have 2 modules
with the same name, and the same symbol, so we will need to address that at
some point.

Also, a new "linkage_name" field has been added to the function and variable
structs, which is the name used when linking. "name" still will reference
the name of the variable as it is used in the program normally.