Fix C backend using linkage name instead name in some places:
To reduce collisions in exported names, name mangling was added. This meant
that there are 2 different names each backend cares about: the normal name
(what the user sees), and the linkage name (unique name, only seen by other
linkers or compilers). These linkage names are not valid C identifiers, and
as such, cannot be used without the use of __asm__ to change the name. Now,
exported vars/funcs in the C backend will use __asm__ to set the proper
linkage name, and the name field will be used for use in expressions (for
example). Eventually we will need to do this for non-exported fields, such as
identifiers which are not valid C identifiers (int, char, etc).