Add shared library for Skull:
Add recipe for building a shared library which wraps up all of the
functionality of Skull in a small (relatively speaking) .so file. The reason
I am doing this so that other programs to easily/dynamically call Skull, and
more specifically, so that I can write a python API for Skull using said
shared library.