Use LLVM API to emit object files instead of external llc command:
This speeds up compilations by about 40ms in my testing.

I also fixed a bug where the LLVM target machine is not actually set, and is
silently optimizing the LLVM anyways. The object emitter needs the target
machine, and is better to be set anyways.

Also add a test for printing assembly to stdout using -o -