Replace SEMANTIC_STATE global variable with function parameter:
Having a single global variable for storing the semantic parser state is
problematic for a few reasons:

* You won't be able to parse multiple files in parallel
* Higher potential for use-after-free errors
* Harder to test

A potential improvement would be to require backends to handle the semantic
state in whatever fasion they see fit: The C backend for example, doesn't
really need to use the semantic state, and shouldn't be required to take a
parameter it will never need.

Since the semantic state is a pointer instead of a struct, there is a potential
for some null pointer-related segfaults, though those should be unlikely, and
very easy to find if they arise.