Fix zero byte files causing segfaults:
Originally the pipeline would be allowed to continue if --werror was not
specified, but there was no test for it, because when --werror isnt
specified, it segfaults. Now it will return early regardless (and, if the
file is empty, there is nothing to compile), but only return an error
if --werror is specified.