Fix functions causing basic blocks to get out of order:
When a function was defined in an if block (or any other block), the function
generating code would take the last basic block, which was not the block that
was set before generating the function. Now we get the basic block currently
selected by the builder, and set that back when we are done.

Also fix other basic block related "issues", despite there not being anything
immediately wrong with it.