Add basic namespace support:
By basic I mean you can only define the namespace itself and you will have no
way to access anything defined in the namespace (yet). There will need to be
extra parsing modifications to allow for member access (dot operator), in
addition to ABI considerations for exported namespaces and namespace members.

Note: I had to remove the const keywords from some of the validate functions,
since the `symbol` field is being accessed directly on the node, though you
are allowed to modify the contents of a field from a pointer.