Add ability to access symbols in namespaces:
Now you have the ability to access symbols inside namespaces using the dot
operator. These are essentially identifiers which are allowed to have dots
in them, and can only be used in expressions, ie, cannot be used in variable,
function, and type aliases.

These dot identifiers cannot be used in import statements yet, since there
would be nothing to import.

There are some namespace features which are not implemented yet:

* Debug info (LLVM Backend)
* Accessing namespace values in the C backend (dont know how to go about it yet)
* Better error handling

There is error handling (and tests!), but you wont get a very helpful error
message when something namespace-related goes awry.