Minor cleanups:
Add errno handling in more places, and remove malloc() when file length is 0.

In testing whether fseek() would fail on character devices (ie, /dev/urandom),
I created a symlink, only to find that the symlinks are "empty" files, and as
such, dont read the underlying file contents. This seems like a bug, but if I
fix it, I will also open up the possibilities to (attempt) to fseek() files
like /dev/urandom. Dont know how/if I want to handle this use case.