Fix certain keywords being able to be used in identifiers:
Since "new identifiers" (ones with a trailing ":", used for explictly typed
variables), the ":" is shaved off, and is re-checked to see if it is a keyword,
and if it is, reject it. There where some keywords which weren't being checked.

In addition, I turned the magic strings into named #define's, so if I ever
decide to change them, I only have to change them in one place (and then the
tests).