Add debug info to the C backend:
I have always been wanting to do this, and it was only after watching [this](https://www.youtube.com/watch?v=ELeZAKCN4tY)
youtube video did I figure out how. Basically, you can use the `#line`
directive which will tell the compiler that the next line is actually at the
above line and filename. This does inflate the size of the emitted C code a
bit, especially if the path of the file being compiled is long. You might be
able to drop subsequent file names so long as the first `#line` contains a file
path, though I have not checked this.

I don't plan on working on Skull very often, this was just a small improvement
that I wanted to expirement with (and I am very pleased with the results).